yearlies: Goal Tree View React UI (first pass)
- Two swimlane rows with synchronized horizontal scroll and drag-to-scroll
- Rubber-band bounce at the future end; peek quarter (Q1 2027) visible
  only in the rubber-band zone, unreachable via nav buttons
- Async Goals window command to fix Win32 deadlock on open
- Compact single-row header with Today/prev/next nav
- Visual details: active-card highlight, edge fade (future side only),
  short month abbreviations, future/done badge labels
- Peek-quarter design decision documented in spec

+++ src/utils/calendar.ts
@@ -20,7 +20,7 @@ export function isCurrentQuarter(q: QuarterDisplay): boolean {
 
 /** Display name for a calendar month, e.g. "May". */
 export function getMonthInfo(month: number, year: number, locale: string): MonthInfo {
-    const label = new Intl.DateTimeFormat(locale, { month: 'long' }).format(
+    const label = new Intl.DateTimeFormat(locale, { month: 'short' }).format(
         new Date(year, month - 1, 1),
     )
     return { month, year, label }
